Fixed random icon shuffling and improved hiding
Hiding moves the window vertically out of the clipping area,
so we're sure icons are not visible.
Sorting now takes place when repainting, after checking the
tray icons sorting does not respect the config files.

+++ trayion/ui.c
@@ -324,14 +324,15 @@ void print_item_info(const char* fname)
 	list_for_each (n, &systray_list) {
 		item = list_entry (n, struct systray_item, systray_list);
 		if (item->rank >= 0){
-			fprintf(fptri, "ID = 0x%x ", item->window_id);
+			fprintf(fptri, "ID = 0x%x ", (unsigned int) item->window_id);
 			if(!skip_free){
 				XFree(class_hint->res_class); 
 				XFree(class_hint->res_name); 
 			}
 			if (XGetClassHint(main_disp, item->window_id, class_hint)){
 				fprintf(fptr, "%s\n", class_hint->res_class);
-				fprintf(fptri, "class: '%s', name: '%s'", class_hint->res_class, class_hint->res_name);
+				fprintf(fptri, "class: '%s', name: '%s'",
+					class_hint->res_class, class_hint->res_name);
 				skip_free = 0;
 			} else {
 				skip_free = 1;
@@ -348,7 +349,9 @@ void print_item_info(const char* fname)
 				XFree(class_hint->res_name); 
 			}
 			XGetClassHint(main_disp, item->window_id, class_hint);
-			fprintf(fptri, "ID = 0x%x class: '%s', name: '%s'\n", item->window_id, class_hint->res_class, class_hint->res_name);
+			fprintf(fptri, "ID = 0x%x class: '%s', name: '%s'\n",
+				(unsigned int) item->window_id,
+				class_hint->res_class, class_hint->res_name);
 			skip_free=0;
 			fprintf(fptr, "%s\n", class_hint->res_class);
 		}
